+++ CHANGELOG.md
@@ -13,6 +13,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 **Other changes:**
 
 * Clean up some console noise on chat, SE.com and other non-QA sites that don't have the same `StackExchange` object as normal QA pages.
+* Add missing `"use strict;"`.  SOUP code should already be strict mode compliant, but this ensures it.
 
 
 1.50 (2018-01-23)