+++ CHANGELOG.md
@@ -46,6 +46,10 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * [mso306552](https://meta.stackoverflow.com/q/306552) was broken by new inline SVG icons, reimplement in pure CSS (with a new custom icon design).
 * Partially rewrote and renamed [boardgames1152](https://boardgames.meta.stackexchange.com/q/1152) as [boardgames1652](https://boardgames.meta.stackexchange.com/q/1652).
 
+**Changes in 1.50.1:**
+
+* Update [mse248646](https://meta.stackexchange.com/q/248646) fix to work with new `.remainingCommentsCount()` method in SE comments UI.
+
 
 1.48 (20 Sep 2017)
 ====